fix(climber): Adjustment should be a double
              Allows for the joystick value to be passed directly

+++ src/main/java/frc/robot/subsystems/climber/ClimberBase.java
@@ -37,8 +37,8 @@ public abstract class ClimberBase extends SubsystemBase {
 
   public abstract void stopWinch();
 
-  public abstract void adjustLeftWinch(int adjust);
-  public abstract void adjustRightWinch(int adjust);
+  public abstract void adjustLeftWinch(double adjust);
+  public abstract void adjustRightWinch(double adjust);
 
   public abstract SparkMaxController getLeftWinch();
